+++ bin/tasks/additions.sh
@@ -10,7 +10,7 @@ chmod 777 ${PROJECT_PATH}/media
 
 echo "${YELLOW}Generating default README.md...${TXTRESET}"
 
-NEWEST_AIR_VERSION="8.4.1"
+NEWEST_AIR_VERSION="8.4.2"
 NEWEST_WORDPRESS_VERSION="5.9.3"
 NEWEST_PHP_VERSION="7.4"
 CURRENT_DATE=$(LC_TIME=en_US date '+%d %b %Y' |tr ' ' '_');