+++ js/src/front-end.js
@@ -7,7 +7,7 @@
 import reframe from 'reframe.js';
 // eslint-disable-next-line no-unused-vars
 import getLocalization from './modules/localization';
-import { styleExternalLinks, getChildAltText } from './modules/external-link';
+import { styleExternalLinks, getChildAltText, initExternalLinkLabels } from './modules/external-link';
 import initAnchors from './modules/anchors';
 import backToTop from './modules/top';
 import 'what-input';
@@ -18,28 +18,13 @@ import './modules/navigation';
 document.body.classList.remove('no-js');
 document.body.classList.add('js');
 
-// Fit video embeds to container
-reframe('.wp-has-aspect-ratio iframe');
-
-// Style external links
-styleExternalLinks();
-
 document.addEventListener('DOMContentLoaded', () => {
   initAnchors();
   backToTop();
-});
+  styleExternalLinks();
+  initExternalLinkLabels();
+  getChildAltText();
 
-// Add aria-labels to links without text or aria-labels and contain image with alt text
-const links = [...document.querySelectorAll('a')];
-// eslint-disable-next-line no-unused-vars
-const linksWithImgChildren = links.forEach((link) => {
-  // If link already has text content or an aria label no need to add aria-label
-  if (link.textContent.trim() !== '' || link.ariaLabel) {
-    return;
-  }
-
-  const ariaLabel = getChildAltText(link);
-  if (ariaLabel !== '') {
-    link.ariaLabel = ariaLabel;
-  }
+  // Fit video embeds to container
+  reframe('.wp-has-aspect-ratio iframe');
 });